refactor: rename files, and remove legacy code
